% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompile.R
\name{precompileRcppNT2}
\alias{precompileRcppNT2}
\title{Pre-compile the RcppNT2 header}
\usage{
precompileRcppNT2()
}
\description{
Pre-compiles the \code{RcppNT2.h} header, for faster compilation. This can be
useful if you need to improve the speed of compilation when iterating with,
for example, \code{Rcpp::sourceCpp()}.
}
\details{
Note that the precompiled header can be quite large (~400MB) so be sure that
you have ample disk space before precompiling the header.
}

