% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-RcppNT2.R
\docType{package}
\name{RcppNT2-package}
\alias{RcppNT2}
\alias{RcppNT2-package}
\title{RcppNT2}
\description{
\code{RcppNT2} provides bindings to the
\href{https://github.com/jfalcou/nt2}{Numerical Template Toolbox}
(\code{NT2}). \code{NT2} is an open source C++ library aimed at
simplifying the development, debugging and optimization
of high-performance computing applications by providing a
\strong{Matlab}-like syntax that eases the transition between
prototype and actual application.
}

